Catalyst integration
Use AMReX from https://github.com/acbauer/amrex with commit
31c355726a7384dbb06ebcac2cbdb23ce242f181. The LyA example has been
tested in serial and parallel and with and without AMR enabled
though testing with AMR enabled didn't actually contain any
refinement. If vtkImageData output is desired can uncomment the
"DEFINES += -DCATALYST_IMAGEDATA" line in GNUmakefile. Nyx then
needs to be run with AMR disabled and the number of MPI ranks
equal to the number of blocks/grids. Need to set INCLUDE_LOCATIONS
and XTRALIBS to have the proper path to ParaView Catalyst. In
AMReX, the path to gridwriter.py (or whatever Catalyst script
is desired) needs to be changed in Src/Amr/AMReX_Amr.cpp.